update meeting screen
